+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 # Refactorings:
 
+* Rename message retrieval events/sagas and use ack feature
 * Rename LOAD_MESSAGES event and incomingMessages reducer
 * Use socket.io ack for CREATE_CHANNEL event
 * Introduce socket.io acknowledgements